Update SQL part 2
DB SQL operations to create dv table and combine with iv table to make a single dataframe. These cases are with cid as the unit of analysis. Can manipulate further df to make market unit of analysis. To make asset-related (e.g., manufacturer) part of analysis, will need to go back to output table.  NOTE: this table too is based on UNIVERSE, not on SAMPLE. Sample will be created with simple LEFT JOIN using cid from cids in sample table.